+++ sharding-jdbc-core/src/main/java/com/dangdang/ddframe/rdb/sharding/parsing/parser/context/OrderItem.java
@@ -41,29 +41,34 @@ public final class OrderItem {
     
     private final OrderType type;
     
+    private final OrderType nullOrderType;
+    
     private int index = -1;
     
     private Optional<String> alias;
     
-    public OrderItem(final String name, final OrderType type, final Optional<String> alias) {
+    public OrderItem(final String name, final OrderType type, final OrderType nullOrderType, final Optional<String> alias) {
         this.owner = Optional.absent();
         this.name = Optional.of(name);
         this.type = type;
+        this.nullOrderType = nullOrderType;
         this.alias = alias;
     }
     
-    public OrderItem(final String owner, final String name, final OrderType type, final Optional<String> alias) {
+    public OrderItem(final String owner, final String name, final OrderType type, final OrderType nullOrderType, final Optional<String> alias) {
         this.owner = Optional.of(owner);
         this.name = Optional.of(name);
         this.type = type;
+        this.nullOrderType = nullOrderType;
         this.alias = alias;
     }
     
-    public OrderItem(final int index, final OrderType type) {
+    public OrderItem(final int index, final OrderType type, final OrderType nullOrderType) {
         owner = Optional.absent();
         name = Optional.absent();
         this.index = index;
         this.type = type;
+        this.nullOrderType = nullOrderType;
         alias = Optional.absent();
     }