+++ sharding-jdbc-core/src/test/java/com/dangdang/ddframe/rdb/sharding/merger/MergeEngineTest.java
@@ -81,7 +81,7 @@ public final class MergeEngineTest {
     
     @Test
     public void assertBuildOrderByStreamResultSetMerger() throws SQLException {
-        selectStatement.getOrderByItems().add(new OrderItem(1, OrderType.DESC));
+        selectStatement.getOrderByItems().add(new OrderItem(1, OrderType.DESC, OrderType.ASC));
         mergeEngine = new MergeEngine(DatabaseType.MySQL, resultSets, selectStatement);
         assertThat(mergeEngine.merge(), instanceOf(OrderByStreamResultSetMerger.class));
     }
@@ -89,7 +89,7 @@ public final class MergeEngineTest {
     @Test
     public void assertBuildOrderByStreamResultSetMergerWithLimit() throws SQLException {
         selectStatement.setLimit(new Limit(true));
-        selectStatement.getOrderByItems().add(new OrderItem(1, OrderType.DESC));
+        selectStatement.getOrderByItems().add(new OrderItem(1, OrderType.DESC, OrderType.ASC));
         mergeEngine = new MergeEngine(DatabaseType.MySQL, resultSets, selectStatement);
         ResultSetMerger actual = mergeEngine.merge();
         assertThat(actual, instanceOf(LimitDecoratorResultSetMerger.class));
@@ -98,8 +98,8 @@ public final class MergeEngineTest {
     
     @Test
     public void assertBuildGroupByStreamResultSetMerger() throws SQLException {
-        selectStatement.getGroupByItems().add(new OrderItem(1, OrderType.DESC));
-        selectStatement.getOrderByItems().add(new OrderItem(1, OrderType.DESC));
+        selectStatement.getGroupByItems().add(new OrderItem(1, OrderType.DESC, OrderType.ASC));
+        selectStatement.getOrderByItems().add(new OrderItem(1, OrderType.DESC, OrderType.ASC));
         mergeEngine = new MergeEngine(DatabaseType.MySQL, resultSets, selectStatement);
         assertThat(mergeEngine.merge(), instanceOf(GroupByStreamResultSetMerger.class));
     }
@@ -107,8 +107,8 @@ public final class MergeEngineTest {
     @Test
     public void assertBuildGroupByStreamResultSetMergerWithLimit() throws SQLException {
         selectStatement.setLimit(new Limit(true));
-        selectStatement.getGroupByItems().add(new OrderItem(1, OrderType.DESC));
-        selectStatement.getOrderByItems().add(new OrderItem(1, OrderType.DESC));
+        selectStatement.getGroupByItems().add(new OrderItem(1, OrderType.DESC, OrderType.ASC));
+        selectStatement.getOrderByItems().add(new OrderItem(1, OrderType.DESC, OrderType.ASC));
         mergeEngine = new MergeEngine(DatabaseType.MySQL, resultSets, selectStatement);
         ResultSetMerger actual = mergeEngine.merge();
         assertThat(actual, instanceOf(LimitDecoratorResultSetMerger.class));
@@ -117,7 +117,7 @@ public final class MergeEngineTest {
     
     @Test
     public void assertBuildGroupByMemoryResultSetMerger() throws SQLException {
-        selectStatement.getGroupByItems().add(new OrderItem(1, OrderType.DESC));
+        selectStatement.getGroupByItems().add(new OrderItem(1, OrderType.DESC, OrderType.ASC));
         mergeEngine = new MergeEngine(DatabaseType.MySQL, resultSets, selectStatement);
         assertThat(mergeEngine.merge(), instanceOf(GroupByMemoryResultSetMerger.class));
     }
@@ -125,7 +125,7 @@ public final class MergeEngineTest {
     @Test
     public void assertBuildGroupByMemoryResultSetMergerWithLimit() throws SQLException {
         selectStatement.setLimit(new Limit(true));
-        selectStatement.getGroupByItems().add(new OrderItem(1, OrderType.DESC));
+        selectStatement.getGroupByItems().add(new OrderItem(1, OrderType.DESC, OrderType.ASC));
         mergeEngine = new MergeEngine(DatabaseType.MySQL, resultSets, selectStatement);
         ResultSetMerger actual = mergeEngine.merge();
         assertThat(actual, instanceOf(LimitDecoratorResultSetMerger.class));