+++ sharding-jdbc-core/src/test/java/com/dangdang/ddframe/rdb/sharding/merger/orderby/OrderByStreamResultSetMergerTest.java
@@ -54,7 +54,7 @@ public final class OrderByStreamResultSetMergerTest {
         when(resultSet.getMetaData()).thenReturn(resultSetMetaData);
         resultSets = Lists.newArrayList(resultSet, mock(ResultSet.class), mock(ResultSet.class));
         selectStatement = new SelectStatement();
-        selectStatement.getOrderByItems().add(new OrderItem(1, OrderType.ASC));
+        selectStatement.getOrderByItems().add(new OrderItem(1, OrderType.ASC, OrderType.ASC));
     }
     
     @Test