+++ sharding-jdbc-core/src/test/java/com/dangdang/ddframe/rdb/sharding/merger/orderby/OrderByValueTest.java
@@ -54,11 +54,11 @@ public final class OrderByValueTest {
     
     @Test
     public void assertCompareToForAsc() throws SQLException {
-        OrderByValue orderByValue1 = new OrderByValue(resultSet1, Arrays.asList(new OrderItem(1, OrderType.ASC), new OrderItem(2, OrderType.ASC)), OrderType.ASC);
+        OrderByValue orderByValue1 = new OrderByValue(resultSet1, Arrays.asList(new OrderItem(1, OrderType.ASC, OrderType.ASC), new OrderItem(2, OrderType.ASC, OrderType.ASC)));
         assertTrue(orderByValue1.next());
         when(resultSet2.getObject(1)).thenReturn("3");
         when(resultSet2.getObject(2)).thenReturn("4");
-        OrderByValue orderByValue2 = new OrderByValue(resultSet2, Arrays.asList(new OrderItem(1, OrderType.ASC), new OrderItem(2, OrderType.ASC)), OrderType.ASC);
+        OrderByValue orderByValue2 = new OrderByValue(resultSet2, Arrays.asList(new OrderItem(1, OrderType.ASC, OrderType.ASC), new OrderItem(2, OrderType.ASC, OrderType.ASC)));
         assertTrue(orderByValue2.next());
         assertTrue(orderByValue1.compareTo(orderByValue2) < 0);
         assertFalse(orderByValue1.getResultSet().next());
@@ -67,11 +67,11 @@ public final class OrderByValueTest {
     
     @Test
     public void assertCompareToForDesc() throws SQLException {
-        OrderByValue orderByValue1 = new OrderByValue(resultSet1, Arrays.asList(new OrderItem(1, OrderType.DESC), new OrderItem(2, OrderType.DESC)), OrderType.ASC);
+        OrderByValue orderByValue1 = new OrderByValue(resultSet1, Arrays.asList(new OrderItem(1, OrderType.DESC, OrderType.ASC), new OrderItem(2, OrderType.DESC, OrderType.ASC)));
         assertTrue(orderByValue1.next());
         when(resultSet2.getObject(1)).thenReturn("3");
         when(resultSet2.getObject(2)).thenReturn("4");
-        OrderByValue orderByValue2 = new OrderByValue(resultSet2, Arrays.asList(new OrderItem(1, OrderType.DESC), new OrderItem(2, OrderType.DESC)), OrderType.ASC);
+        OrderByValue orderByValue2 = new OrderByValue(resultSet2, Arrays.asList(new OrderItem(1, OrderType.DESC, OrderType.ASC), new OrderItem(2, OrderType.DESC, OrderType.ASC)));
         assertTrue(orderByValue2.next());
         assertTrue(orderByValue1.compareTo(orderByValue2) > 0);
         assertFalse(orderByValue1.getResultSet().next());
@@ -80,11 +80,11 @@ public final class OrderByValueTest {
     
     @Test
     public void assertCompareToWhenEqual() throws SQLException {
-        OrderByValue orderByValue1 = new OrderByValue(resultSet1, Arrays.asList(new OrderItem(1, OrderType.ASC), new OrderItem(2, OrderType.DESC)), OrderType.ASC);
+        OrderByValue orderByValue1 = new OrderByValue(resultSet1, Arrays.asList(new OrderItem(1, OrderType.ASC, OrderType.ASC), new OrderItem(2, OrderType.DESC, OrderType.ASC)));
         assertTrue(orderByValue1.next());
         when(resultSet2.getObject(1)).thenReturn("1");
         when(resultSet2.getObject(2)).thenReturn("2");
-        OrderByValue orderByValue2 = new OrderByValue(resultSet2, Arrays.asList(new OrderItem(1, OrderType.ASC), new OrderItem(2, OrderType.DESC)), OrderType.ASC);
+        OrderByValue orderByValue2 = new OrderByValue(resultSet2, Arrays.asList(new OrderItem(1, OrderType.ASC, OrderType.ASC), new OrderItem(2, OrderType.DESC, OrderType.ASC)));
         assertTrue(orderByValue2.next());
         assertThat(orderByValue1.compareTo(orderByValue2), is(0));
         assertFalse(orderByValue1.getResultSet().next());