+++ sharding-jdbc-core/src/test/java/com/dangdang/ddframe/rdb/sharding/parsing/parser/base/AbstractBaseParseSQLTest.java
@@ -131,7 +131,8 @@ public abstract class AbstractBaseParseSQLTest extends AbstractBaseParseTest {
         Iterator<OrderItem> orderByColumns = getExpectedOrderByColumns().iterator();
         for (OrderItem each : actual.getOrderByItems()) {
             OrderItem expectedOrderItem = orderByColumns.next();
-            assertTrue(new ReflectionEquals(expectedOrderItem).matches(each));
+            // TODO assert nullOrderType
+            assertTrue(new ReflectionEquals(expectedOrderItem, "nullOrderType").matches(each));
         }
         assertFalse(orderByColumns.hasNext());
     }
@@ -140,7 +141,8 @@ public abstract class AbstractBaseParseSQLTest extends AbstractBaseParseTest {
         Iterator<OrderItem> groupByColumns = getExpectedGroupByColumns().iterator();
         for (OrderItem each : actual.getGroupByItems()) {
             OrderItem groupByColumn = groupByColumns.next();
-            assertTrue(new ReflectionEquals(groupByColumn).matches(each));
+            // TODO assert nullOrderType
+            assertTrue(new ReflectionEquals(groupByColumn, "nullOrderType").matches(each));
         }
         assertFalse(groupByColumns.hasNext());
     }